changed theme for code snippet and fixed responsivness on ultrawide for videos
