+++ src/components/Cursor.jsx
@@ -1,0 +1,31 @@
+import React, { useRef, useEffect } from "react";
+
+export default function Cursor() {
+  const blobRef = useRef(null);
+
+  useEffect(() => {
+    const handlePointerMove = (event) => {
+      const { clientX, clientY } = event;
+
+      blobRef.current.animate(
+        {
+          left: `${clientX}px`,
+          top: `${clientY}px`,
+        },
+        { duration: 3000, fill: "forwards" }
+      );
+    };
+
+    window.addEventListener("pointermove", handlePointerMove);
+
+    return () => {
+      window.removeEventListener("pointermove", handlePointerMove);
+    };
+  }, []);
+  return (
+    <>
+      <div className="blob" ref={blobRef}></div>
+      <div className="blur"></div>
+    </>
+  );
+}
