+++ src/index.css
@@ -119,6 +119,36 @@ main {
   overflow: hidden;
 }
 
+.blob {
+  background-color: white;
+  height: 25vmax;
+  aspect-ratio: 1;
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  translate: -50% -50%;
+  border-radius: 50%;
+  background: linear-gradient(to right, var(--accent), var(--primary));
+  animation: rotate 20s infinite;
+  opacity: 0.6;
+  padding: 4rem;
+}
+
+.blur {
+  height: 110%;
+  width: 110%;
+  position: absolute;
+  backdrop-filter: blur(12vmax);
+  left: 0px;
+}
+
+pre {
+  border: unset !important;
+  background: unset !important;
+  padding: 0rem !important;
+  margin: 1rem !important;
+}
+
 .burger-icon {
   display: flex;
   flex-direction: column;
