+++ src/sections/AboutMe/AboutMe.module.css
@@ -210,7 +210,7 @@ video {
   width: 100%;
   height: 100%;
   border-radius: 10px;
-  max-width: 585px;
+  max-width: 435px;
 }
 
 .linkContainer {
