+++ src/sections/Hero/Hero.module.css
@@ -7,29 +7,6 @@
   border-radius: 10px;
 }
 
-.blob {
-  background-color: white;
-  height: 25vmax;
-  aspect-ratio: 1;
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  translate: -50% -50%;
-  border-radius: 50%;
-  background: linear-gradient(to right, var(--accent), var(--primary));
-  animation: rotate 20s infinite;
-  opacity: 0.6;
-  padding: 4rem;
-}
-
-.blur {
-  height: 110%;
-  width: 110%;
-  position: absolute;
-  backdrop-filter: blur(12vmax);
-  left: 0px;
-}
-
 @keyframes rotate {
   from {
     rotate: 0deg;
