+++ src/sections/Projects/ProjectCard.jsx
@@ -17,9 +17,6 @@ export default function ProjectCard(props) {
     e.currentTarget.classList.remove(styles.hovered);
   }
 
-  console.log(window.innerHeight);
-  console.log(window.innerWidth);
-
   const projectCardArr = projects.map((project, index) => {
     const isActive = selectedProject == index;
